+++ src/main/java/com/codeborne/selenide/impl/DownloadFileWithCdp.java
@@ -13,10 +13,10 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.devtools.DevTools;
 import org.openqa.selenium.devtools.HasDevTools;
-import org.openqa.selenium.devtools.v132.browser.Browser;
-import org.openqa.selenium.devtools.v132.browser.model.DownloadProgress;
-import org.openqa.selenium.devtools.v132.browser.model.DownloadWillBegin;
-import org.openqa.selenium.devtools.v132.page.Page;
+import org.openqa.selenium.devtools.v135.browser.Browser;
+import org.openqa.selenium.devtools.v135.browser.model.DownloadProgress;
+import org.openqa.selenium.devtools.v135.browser.model.DownloadWillBegin;
+import org.openqa.selenium.devtools.v135.page.Page;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,8 +32,8 @@ import static com.codeborne.selenide.impl.WebdriverUnwrapper.cast;
 import static java.lang.System.currentTimeMillis;
 import static java.util.Collections.emptyMap;
 import static java.util.Objects.requireNonNull;
-import static org.openqa.selenium.devtools.v132.browser.Browser.downloadProgress;
-import static org.openqa.selenium.devtools.v132.browser.Browser.downloadWillBegin;
+import static org.openqa.selenium.devtools.v135.browser.Browser.downloadProgress;
+import static org.openqa.selenium.devtools.v135.browser.Browser.downloadWillBegin;
 
 public class DownloadFileWithCdp {
   private static final Logger log = LoggerFactory.getLogger(DownloadFileWithCdp.class);
@@ -127,7 +127,7 @@ public class DownloadFileWithCdp {
     if (cdpBrowser.isPresent() && isChromium(webDriver)) {
       DevTools devTools = cdpBrowser.get().getDevTools();
       devTools.createSessionIfThereIsNotOne();
-      devTools.send(Page.enable());
+      devTools.send(Page.enable(Optional.empty()));
       return devTools;
     } else {
       throw new IllegalArgumentException(