+++ src/test/java/com/codeborne/selenide/impl/ElementFinderTest.java
@@ -12,28 +12,28 @@ final class ElementFinderTest {
 
   @Test
   void toStringForFinderByCssSelectors() {
-    ElementFinder parent = new ElementFinder(driver, null, By.tagName("table"), 0);
-    assertThat(new ElementFinder(driver, null, By.id("app"), 0))
+    ElementFinder parent = new ElementFinder(driver, null, By.tagName("table"), 0, false);
+    assertThat(new ElementFinder(driver, null, By.id("app"), 0, false))
       .hasToString("{#app}");
-    assertThat(new ElementFinder(driver, null, By.id("app"), 3))
+    assertThat(new ElementFinder(driver, null, By.id("app"), 3, false))
       .hasToString("{#app[3]}");
-    assertThat(new ElementFinder(driver, parent, By.id("app"), 0))
+    assertThat(new ElementFinder(driver, parent, By.id("app"), 0, false))
       .hasToString("{table/#app}");
-    assertThat(new ElementFinder(driver, parent, By.id("app"), 3))
+    assertThat(new ElementFinder(driver, parent, By.id("app"), 3, false))
       .hasToString("{table/#app[3]}");
   }
 
   @Test
   void toStringForFinderByXpathExpiration() {
-    ElementFinder parent = new ElementFinder(driver, null, By.xpath("//table"), 0);
+    ElementFinder parent = new ElementFinder(driver, null, By.xpath("//table"), 0, false);
 
-    assertThat(new ElementFinder(driver, null, By.xpath("//*[@id='app']"), 0))
+    assertThat(new ElementFinder(driver, null, By.xpath("//*[@id='app']"), 0, false))
       .hasToString("{By.xpath: //*[@id='app']}");
-    assertThat(new ElementFinder(driver, null, By.xpath("//*[@id='app']"), 3))
+    assertThat(new ElementFinder(driver, null, By.xpath("//*[@id='app']"), 3, false))
       .hasToString("{By.xpath: //*[@id='app'][3]}");
-    assertThat(new ElementFinder(driver, parent, By.xpath("//*[@id='app']"), 0))
+    assertThat(new ElementFinder(driver, parent, By.xpath("//*[@id='app']"), 0, false))
       .hasToString("{By.xpath: //table/By.xpath: //*[@id='app']}");
-    assertThat(new ElementFinder(driver, parent, By.xpath("//*[@id='app']"), 3))
+    assertThat(new ElementFinder(driver, parent, By.xpath("//*[@id='app']"), 3, false))
       .hasToString("{By.xpath: //table/By.xpath: //*[@id='app'][3]}");
   }
 }